Major refactor

+++ src/user/controller/delete.rs
@@ -1,8 +1,11 @@
 use axum::{extract::State, routing::delete, Router};
 
 use crate::{
-    prisma::user::Data,
-    user::request::{delete::DeleteUserRequest, loggedin::LoggedInUserWithPassword},
+    user::model::{
+        delete::DeleteUserRequest,
+        loggedin::LoggedInUserWithPassword,
+        select::{User, UserSelect},
+    },
     validation::validation_message,
     web::Web,
     GlobalState, WebResult,
@@ -10,23 +13,28 @@ use crate::{
 
 pub fn delete_user() -> Router<GlobalState> {
     async fn delete_user_handler(
-        State(GlobalState { user_service, .. }): State<GlobalState>,
-        LoggedInUserWithPassword(Data {
+        State(GlobalState {
+            user_service,
+            folder_service,
+            ..
+        }): State<GlobalState>,
+        LoggedInUserWithPassword(User {
             id: user_id,
             password: user_password,
             ..
         }): LoggedInUserWithPassword,
         DeleteUserRequest { password, .. }: DeleteUserRequest,
     ) -> WebResult {
-        match user_password == password {
-            true => {
-                user_service.delete_user(user_id).await?;
-                Ok(Web::ok("Deleted user successfully", ()))
-            }
-            false => {
-                Err(validation_message("Password provided does not match current password").into())
-            }
+        if user_password != password {
+            return Err(
+                validation_message("Password provided does not match current password").into(),
+            );
         }
+
+        let UserSelect { id: owner_id, .. } = user_service.delete_user(user_id).await?;
+        let deleted_folder = folder_service.delete_root_folder(owner_id).await?;
+
+        Ok(Web::ok("Deleted user successfully", ()))
     }
     Router::new().route("/delete", delete(delete_user_handler))
 }
